+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 ### Changelog
 
-* 1/7/2016: Added `uname` - [@dblock](https://github.com/dblock).
+* 1/7/2016: Added `uname` and `whoami` - [@dblock](https://github.com/dblock).
 * 1/6/2016: Initial public release with `help`, `ls`, `pwd`, `echo`, `mkdir`, `rmdir`, `touch`, `rm` and `cat` - [@dblock](https://github.com/dblock).